types: Moved a bunch of Stage structs and enums from core to types
