+++ core/src/avm1/object/stage_object.rs
@@ -861,7 +861,7 @@ fn high_quality<'gc>(
     activation: &mut Activation<'_, 'gc, '_>,
     _this: DisplayObject<'gc>,
 ) -> Value<'gc> {
-    use crate::display_object::StageQuality;
+    use ruffle_types::display_object::stage::StageQuality;
     let quality = match activation.context.stage.quality() {
         StageQuality::Best => 2,
         StageQuality::High => 1,
@@ -875,7 +875,7 @@ fn set_high_quality<'gc>(
     _this: DisplayObject<'gc>,
     val: Value<'gc>,
 ) -> Result<(), Error<'gc>> {
-    use crate::display_object::StageQuality;
+    use ruffle_types::display_object::stage::StageQuality;
     let val = val.coerce_to_f64(activation)?;
     if !val.is_nan() {
         // 0 -> Low, 1 -> High, 2 -> Best, but with some odd rules for non-integers.
