+++ desktop/src/main.rs
@@ -18,9 +18,10 @@ use crate::executor::GlutinAsyncExecutor;
 use clap::Parser;
 use isahc::{config::RedirectPolicy, prelude::*, HttpClient};
 use rfd::FileDialog;
-use ruffle_core::{config::Letterbox, Player, PlayerBuilder, StageDisplayState};
+use ruffle_core::{config::Letterbox, Player, PlayerBuilder};
 use ruffle_render_wgpu::clap::{GraphicsBackend, PowerPreference};
 use ruffle_render_wgpu::WgpuRenderBackend;
+use ruffle_types::display_object::stage::StageDisplayState;
 use ruffle_types::events::{KeyCode, PlayerEvent};
 use ruffle_types::tag_utils::SwfMovie;
 use std::io::Read;
