+++ types/Cargo.toml
@@ -22,6 +22,7 @@ generational-arena = "0.2.8"
 instant = "0.1"
 thiserror = "1.0"
 indexmap = "1.9.1"
+bitflags = "1.3.2"
 
 [dependencies.jpeg-decoder]
 version = "0.2.6"
